requirements.txt
opencv-python
opencv-contrib-python
numpy
pandas
flask
mysql-connector-python
pillow
matplotlib